backend/requirements.txt
flask==2.3.3
Flask-Cors==3.0.10
python-dotenv==1.0.0
google-generativeai==0.2.2